misc/HtmlTestSuite/results/test-0505.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>CSS Grid: grid gutters and tracks.</title>
<link href="mailto:svillar@igalia.com" rel="author" title="Sergio Villar"/>
<link href="https://drafts.csswg.org/css-grid-1/#gutters" rel="help"/>
<meta content="Check that gutters contribute to the size of the grid containers and spanning items, and do not alter grid items positioning, margin computation and track sizing." name="assert"/>
<link href="/css/support/grid.css" rel="stylesheet"/>
<link href="/css/support/width-keyword-classes.css" rel="stylesheet"/>
<link href="/fonts/ahem.css" rel="stylesheet" type="text/css"/>
<style>
body { margin: 0px; }

.grid { padding: 3px 7px 5px 1px; border: solid yellow 1px; }

.normalGap { grid-gap: normal; }

.gridGap { grid-gap: 16px; }

.gridRowColumnGaps {
    grid-row-gap: 12px;
    grid-column-gap: 23px;
}

.gridMultipleCols { grid-template-columns: 30px minmax(10px, 50px) 80px; }
.gridMultipleRows { grid-template-rows: 90px 70px min-content; }
.gridAutoAuto { grid-template: auto auto / auto auto; }
.gridMultipleFixed { grid-template: [first] 37px [foo] 57px [bar] 77px [last] / [first] 15px [foo] 25px [bar] 35px [last]; }
.gridFixed100 { grid-template: repeat(2, 100px) / repeat(2, 100px); }
.gridWithPercent { grid-template-columns: 10px 20% repeat(2, 30px); }
.gridWithDoublePercent { grid-template-columns: 60% 40%; }
.gridWithRowPercent { grid-template: 10px 20% 30px / 20px; }
.gridWithRowDoublePercent { grid-template: 60% 40% / 20px; }

.width220 { width: 220px; }
.height100 { height: 100px; }
.gridAutoRows20 { grid-auto-rows: 20px; }

.thirdRowThirdColumn { grid-area: 3 / 3; }
.firstRowThirdColumn { grid-area: 1 / 3; }

div.grid &gt; div { font: 10px/1 Ahem; }

.gridItemMargins {
    margin: 20px 30px 40px 50px;
    width: 20px;
    height: 40px;
}

</style>

<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>
<script type="text/javascript">
    setup({ explicit_done: true });
</script>

</head><body onload="document.fonts.ready.then(() =&gt; { checkLayout('.grid'); })">

<!-- Check that gutters contribute to the size of the grid containers. -->

<div style="position: relative;">
    <div class="grid normalGap gridMultipleCols fit-content" data-expected-height="10" data-expected-width="170"/>
    <div class="grid normalGap gridMultipleRows" data-expected-height="170" data-expected-width="410" style="width: 400px"/>
</div>

<div style="position: relative">
    <div class="grid gridRowColumnGaps fit-content" data-expected-height="62" data-expected-width="103">
        <div class="firstRowFirstColumn" data-expected-height="10" data-expected-width="20" data-offset-x="2" data-offset-y="4">XX</div>
        <div class="secondRowSecondColumn" data-expected-height="30" data-expected-width="50" data-offset-x="45" data-offset-y="26">XX<br/>X<br/>XX XX</div>
    </div>
</div>

<div style="position: relative">
    <div class="grid gridMultipleCols gridRowColumnGaps fit-content" data-expected-height="94" data-expected-width="216">
        <div class="firstRowFirstColumn" data-expected-height="10" data-expected-width="30" data-offset-x="2" data-offset-y="4">XX</div>
        <div class="secondRowSecondColumn" data-expected-height="20" data-expected-width="50" data-offset-x="55" data-offset-y="26">XX<br/>XX XX</div>
        <div class="thirdRowThirdColumn" data-expected-height="30" data-expected-width="80" data-offset-x="128" data-offset-y="58">XXXX XX<br/>X<br/>XX XX</div>
    </div>
</div>

<div style="position: relative">
    <div class="grid gridMultipleRows gridRowColumnGaps fit-content" data-expected-height="224" data-expected-width="196">
        <div class="firstRowFirstColumn" data-expected-height="90" data-expected-width="20" data-offset-x="2" data-offset-y="4">XX</div>
        <div class="secondRowSecondColumn" data-expected-height="70" data-expected-width="50" data-offset-x="45" data-offset-y="106">XX<br/>XX XX</div>
        <div class="thirdRowThirdColumn" data-expected-height="30" data-expected-width="70" data-offset-x="118" data-offset-y="188">XXXX XX<br/>X<br/>XX XX</div>
    </div>
</div>

<!-- Check that gutters do not alter grid items positioning. -->
<div style="position: relative">
    <div class="grid gridMultipleFixed gridRowColumnGaps">
        <div class="firstRowFirstColumn" data-expected-height="37" data-expected-width="15" data-offset-x="2" data-offset-y="4">X X X</div>
        <div class="secondRowSecondColumn" data-expected-height="57" data-expected-width="25" data-offset-x="40" data-offset-y="53">X X</div>
        <div class="thirdRowThirdColumn" data-expected-height="77" data-expected-width="35" data-offset-x="88" data-offset-y="122">XXX XX X XX XX</div>
    </div>
</div>

<div style="position: relative">
    <div class="grid gridMultipleFixed gridRowColumnGaps">
        <div data-expected-height="57" data-expected-width="35" data-offset-x="88" data-offset-y="53" style="grid-row: 2; grid-column: -2 / -1;">X X X</div>
        <div data-expected-height="106" data-expected-width="35" data-offset-x="88" data-offset-y="4" style="grid-row: 1 / bar; grid-column: bar">X XX X XX XX XX X</div>
        <div data-expected-height="77" data-expected-width="15" data-offset-x="2" data-offset-y="122" style="grid-row: -2; grid-column-end: foo">X X</div>
    </div>
</div>

<!-- Check that gutters do not alter track sizing. -->
<div style="position: relative">
    <div class="grid gridRowColumnGaps fit-content" data-expected-height="94" data-expected-width="166">
        <div class="firstRowFirstColumn" data-expected-height="10" data-expected-width="20" data-offset-x="2" data-offset-y="4">XX</div>
        <div class="secondRowSecondColumn" data-expected-height="20" data-expected-width="30" data-offset-x="45" data-offset-y="26">X X<br/>X X</div>
        <div class="thirdRowThirdColumn" data-expected-height="30" data-expected-width="60" data-offset-x="98" data-offset-y="58">XXX XX<br/>X<br/>XX XX</div>
    </div>
</div>

<div style="position: relative">
    <div class="grid gridRowColumnGaps gridMultipleCols gridMultipleRows fit-content" data-expected-height="224" data-expected-width="216">
        <div class="firstRowFirstColumn" data-expected-height="90" data-expected-width="30" data-offset-x="2" data-offset-y="4">XX</div>
        <div class="secondRowSecondColumn" data-expected-height="70" data-expected-width="50" data-offset-x="55" data-offset-y="106">X X<br/>X X</div>
        <div class="thirdRowThirdColumn" data-expected-height="30" data-expected-width="80" data-offset-x="128" data-offset-y="188">XXX XX<br/>X<br/>XX XX</div>
    </div>
</div>

<div style="position: relative">
    <div class="grid gridWithPercent width220 gridRowColumnGaps gridAutoRows20" data-expected-height="94" data-expected-width="230">
        <div class="firstRowFirstColumn" data-expected-height="20" data-expected-width="10" data-offset-x="2" data-offset-y="4"/>
        <div class="secondRowSecondColumn" data-expected-height="20" data-expected-width="44" data-offset-x="35" data-offset-y="36"/>
        <div class="thirdRowThirdColumn" data-expected-height="20" data-expected-width="30" data-offset-x="102" data-offset-y="68"/>
    </div>
</div>

<div style="position: relative">
    <div class="grid gridWithRowPercent gridRowColumnGaps width220 height100" data-expected-height="110" data-expected-width="230">
        <div class="firstRowFirstColumn" data-expected-height="10" data-expected-width="20" data-offset-x="2" data-offset-y="4"/>
        <div class="secondRowFirstColumn" data-expected-height="20" data-expected-width="20" data-offset-x="2" data-offset-y="26"/>
        <div class="thirdRowAutoColumn" data-expected-height="30" data-expected-width="20" data-offset-x="2" data-offset-y="58"/>
    </div>
</div>

<!-- Check that gutters contribute to the size of spanning items. -->
<div style="position: relative">
    <div class="grid gridGap gridAutoAuto constrainedContainer">
        <div class="secondRowBothColumn" data-expected-height="10" data-expected-width="50" data-offset-x="2" data-offset-y="30">XXXXX</div>
        <div class="firstRowFirstColumn" data-expected-height="10" data-expected-width="17" data-offset-x="2" data-offset-y="4">X</div>
        <div class="firstRowSecondColumn" data-expected-height="10" data-expected-width="17" data-offset-x="35" data-offset-y="4">X</div>
    </div>
</div>

<div style="position: relative">
    <div class="grid gridMultipleFixed gridRowColumnGaps">
        <div data-expected-height="57" data-expected-width="121" data-offset-x="2" data-offset-y="53" style="grid-row: 2; grid-column: 1 / -1;">XXXX X XXXX</div>
        <div data-expected-height="195" data-expected-width="25" data-offset-x="40" data-offset-y="4" style="grid-row: first / last; grid-column-start: 2">X XX X XX X</div>
        <div data-expected-height="106" data-expected-width="63" data-offset-x="2" data-offset-y="4" style="grid-row: 1 / 3; grid-column: first / bar">XXX XX<br/>XX<br/>XXXXX</div>
    </div>
</div>

<div style="position: relative">
    <div class="grid gridWithDoublePercent width220 gridRowColumnGaps gridAutoRows20" data-expected-height="62" data-expected-width="230">
        <div class="firstRowFirstColumn sizedToGridArea" data-expected-height="20" data-expected-width="132" data-offset-x="2" data-offset-y="4"/>
        <div class="secondRowSecondColumn sizedToGridArea" data-expected-height="20" data-expected-width="88" data-offset-x="157" data-offset-y="36"/>
        <div class="secondRowBothColumn sizedToGridArea" data-expected-height="20" data-expected-width="243" data-offset-x="2" data-offset-y="36"/>
    </div>
</div>

<div style="position: relative">
    <div class="grid gridWithRowDoublePercent gridRowColumnGaps width220 height100" data-expected-height="110" data-expected-width="230">
        <div class="firstRowFirstColumn" data-expected-height="60" data-expected-width="20" data-offset-x="2" data-offset-y="4"/>
        <div class="secondRowFirstColumn" data-expected-height="40" data-expected-width="20" data-offset-x="2" data-offset-y="76"/>
        <div class="bothRowFirstColumn" data-expected-height="112" data-expected-width="20" data-offset-x="2" data-offset-y="4"/>
    </div>
</div>

<!-- Check that gutters do not interfere with margins computation. -->
<div style="position: relative">
    <div class="grid gridFixed100 gridGap">
        <div class="gridItemMargins firstRowFirstColumn" data-expected-height="40" data-expected-width="20" data-offset-x="52" data-offset-y="24"/>
        <div class="gridItemMargins firstRowSecondColumn" data-expected-height="40" data-expected-width="20" data-offset-x="168" data-offset-y="24"/>
        <div class="gridItemMargins secondRowFirstColumn" data-expected-height="40" data-expected-width="20" data-offset-x="52" data-offset-y="140"/>
        <div class="gridItemMargins secondRowSecondColumn" data-expected-height="40" data-expected-width="20" data-offset-x="168" data-offset-y="140"/>
    </div>
</div>

<div style="position: relative">
    <div class="grid gridFixed100 verticalRL directionRTL gridGap">
        <div class="gridItemMargins firstRowFirstColumn" data-expected-height="40" data-expected-width="20" data-offset-x="168" data-offset-y="140"/>
        <div class="gridItemMargins firstRowSecondColumn" data-expected-height="40" data-expected-width="20" data-offset-x="168" data-offset-y="24"/>
        <div class="gridItemMargins secondRowFirstColumn" data-expected-height="40" data-expected-width="20" data-offset-x="52" data-offset-y="140"/>
        <div class="gridItemMargins secondRowSecondColumn" data-expected-height="40" data-expected-width="20" data-offset-x="52" data-offset-y="24"/>
    </div>
</div>

</body>
</html>
